requirements.txt
Django>=5.0,<5.1
gunicorn
# .env file management 
python-decouple
# Database
psycopg[binary]
dj-database-url
requests
# for static files in production
whitenoise
# Allauth
django-allauth[socialaccount]
django-allauth-ui
django-widget-tweaks
slippers
stripe